chore(engine): fix clippy errors across gnitz-engine
Mechanical / unambiguous fixes:
- `vec![…; N]` → `[…; N]` for stack-only locals
- `as u64` on a `u64`, raw-pointer self-cast: dropped
- manual nul-terminated byte string → `c"…"` literal
- `% 8 == 0` → `is_multiple_of(8)`
- `loop { match try_consume(…) { Some(…) => …, None => break } }` →
  `while let Some(…) = try_consume(…)`
- `for i in 0..n { v[i] = … }` → `for (i, slot) in v.iter_mut().enumerate().take(n)`
- `iter().copied().collect::<Vec<_>>()` on a `Vec` → `.to_vec()`
- `.clone()` for a single-element slice → `std::slice::from_ref(&x)`
- `async move { fut.await }` / single-poll async block → direct future
- `fn …() -> impl Future + 'a { async move { … } }` → `async fn`
- `let _ = void_returning_call(…)` → bare call
- nested `if`/`if` collapsed via `&&`

Scoped `#[allow]`s for intentional patterns:
- `clippy::approx_constant` on tests using 3.14/2.718 as deliberate
  f32/f64 round-trip fixtures (not PI/E approximations).
- `clippy::erasing_op` + `clippy::identity_op` on expr/tests/batch_tests
  where `0 * MORSEL` / `1 * MORSEL` document register-bank offsets.
- `clippy::too_many_arguments` on three test-only helper signatures
  (`append_row_simple`, `sal_write_group`, `encode_wire`).
- `clippy::type_complexity` on a 7-tuple test-case array literal.

+++ crates/gnitz-engine/src/catalog/tests/index_tests.rs
@@ -15,7 +15,7 @@ fn test_index_creation_and_backfill() {
     let mut bb = BatchBuilder::new(schema);
     for i in 0..5u64 {
         bb.begin_row(i as u128, 1);
-        bb.put_u64((i * 100) as u64);
+        bb.put_u64(i * 100);
         bb.end_row();
     }
     let batch = bb.finish();
